requirements.txt
Flask
Flask-RESTful
Flask-Migrate
Flask-SQLAlchemy
marshmallow
python-dotenv
flasgger